+++ data/src/main/java/com/yapp/bol/data/model/member/MemberListResponse.kt
@@ -1,7 +1,5 @@
 package com.yapp.bol.data.model.member
 
-import com.yapp.bol.data.model.group.MemberDTO
-
 data class MemberListResponse(
     val contents: List<MemberDTO>,
     val cursor: String,
